create ConfigManager utility
this util abstracts the complexity of reading config values that may be set as environment variables, or pulled from a local config file.

the idea is that we should be able to just ask for a config value without having to worry about the execution context.

as long as the config values have been set correctly, we should be able to just ask for a config value and get it. The code doesn't concern itself whether it comes from the environment (when running in lambda) or a local config file.